refactor(toolinteract): fold the 42-file split back into 16 real units
All 42 source files here were created by ONE commit (2920afa9), so the
interact_* filename prefixes are that commit's naming taste — "filenames agree
with history" is one fact counted twice. The independent signal is the call
graph, and it says this package is held together by method pinning: 28 of the
42 files are nothing but methods on *InteractActionHandler, and Go only allows
those in the package that declares the type. No amount of file-shuffling
splits them; only folding does.

Where the graph disagreed with the filenames:
- interact_tracking.go had exactly one caller, HandleBrowserActionSwitchTabImpl.
  It is a private continuation of switch_tab, not a tracking subsystem — it now
  lives in interact_browser.go.
- interact_workflow_types.go was not "types": it holds workflowResult, the
  response envelope every workflow ends in, so no workflow could be read
  without it.
- interact_dom_{params,validation}.go are only ever entered through
  HandleDOMPrimitive; they were never a reusable validation layer.
- The evidence and retry-contract clusters were 5 and 3 files around one mutex
  and one map each. They fold rather than split, for the same reason
  internal/queries went 15 -> 8: exporting a mutex to move its callers is worse
  than keeping the locking discipline in one readable file.

Behavior preservation, proved mechanically:
- The multiset of non-comment code lines (AST-reprinted, comments stripped) is
  identical before and after except for de-duplicated import lines. Zero logic
  lines added, removed or changed.
- `go doc -all` yields the same 90 exported declarations, byte for byte.
- Package coverage unchanged at 49.9%.

cmd/browser-agent/internal/toolinteract: 42 -> 16 source files.
