+++ docs/features/feature/file-upload/index.md
@@ -4,9 +4,8 @@ feature_id: feature-file-upload
 status: shipped
 feature_type: feature
 owners: []
-last_reviewed: 2026-07-26
+last_reviewed: 2026-07-27
 code_paths:
-  - cmd/browser-agent/internal/toolinteract/interact_upload_handler.go
   - cmd/browser-agent/internal/toolinteract/interact_upload.go
   - cmd/browser-agent/upload_handlers.go
   - internal/upload/handlers.go
